[Patchset 3] Add StringView class and switch trace processor to use it
In the TraceProcessor most of the strings derived from
protos are not terminated. Instead of passing a pointer and
a length, just pass a StringView.

Test: perfetto_unittests --gtest_filter=StringViewTest.*

+++ src/base/string_view_unittest.cc
@@ -55,7 +55,7 @@ TEST(StringViewTest, BasicCases) {
 }
 
 TEST(StringViewTest, HashCollisions) {
-  std::unordered_map<size_t, StringView> hashes;
+  std::unordered_map<uint64_t, StringView> hashes;
   std::unordered_set<StringView> sv_set;
   auto insert_sv = [&hashes, &sv_set](StringView sv) {
     hashes.emplace(sv.Hash(), sv);
